services/ingest/requirements.txt
psycopg2-binary
requests
beautifulsoup4
qdrant-client
sentence-transformers
tqdm